........
  r5535 | jfranks | 2007-02-23 05:01:32 -0600 (Fri, 23 Feb 2007) | 8 lines
  	* client-src/sendbackup.c client-src/sendsize.c client-src/sendbackup-star.c
  	  configure.in server-src/amindexd.c server-src/taper.c server-src/diskfile.c
  	  config/acinclude.m4i common-src/amanda.h common-src/util.c
  	  common-src/security-util.c:
  		Fix compilation problems using GCC on solaris.
........


git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@363 a8d146d6-cc15-0410-8900-af154a0219e0

+++ client-src/sendsize.c
@@ -890,7 +890,7 @@ generic_calc_estimates(
     amfree(match_expr);
 
     dbprintf(_("waiting for %s %s child (pid=%d)\n"),
-	      my_argv[0], est->qamdevice, calcpid);
+	      my_argv[0], est->qamdevice, (int)calcpid);
     waitpid(calcpid, &wait_status, 0);
     if (WIFSIGNALED(wait_status)) {
 	errmsg = vstrallocf(_("%s terminated with signal %d: see %s"),
@@ -910,7 +910,7 @@ generic_calc_estimates(
     }
     dbprintf(_("after %s %s wait: child pid=%d status=%d\n"),
 	      my_argv[0], est->qamdevice,
-	      calcpid, WEXITSTATUS(wait_status));
+	      (int)calcpid, WEXITSTATUS(wait_status));
 
     dbprintf(_(".....\n"));
     dbprintf(_("estimate time for %s: %s\n"),
