+++ common-src/amanda.h
@@ -789,15 +789,15 @@ void	areads_relbuf(int fd);
 #define	NUM_STR_SIZE	128		/* a generic number buffer size */
 
 #define	skip_whitespace(ptr,c) do {					\
-    while((c) != '\n' && isspace(c)) (c) = *(ptr)++;			\
+    while((c) != '\n' && isspace((int)c)) (c) = *(ptr)++;		\
 } while(0)
 
 #define	skip_non_whitespace(ptr,c) do {					\
-    while((c) != '\0' && !isspace(c)) (c) = *(ptr)++;			\
+    while((c) != '\0' && !isspace((int)c)) (c) = *(ptr)++;		\
 } while(0)
 
 #define	skip_non_whitespace_cs(ptr,c) do {				\
-    while((c) != '\0' && (c) != '#' && !isspace(c)) (c) = *(ptr)++;	\
+    while((c) != '\0' && (c) != '#' && !isspace((int)c)) (c) = *(ptr)++;\
 } while(0)
 
 #define	skip_non_integer(ptr,c) do {					\
@@ -811,7 +811,7 @@ void	areads_relbuf(int fd);
 
 #define skip_quoted_string(ptr, c) do {					\
     int	iq = 0;								\
-    while (((c) != '\0') && !((iq == 0) && isspace(c))) {		\
+    while (((c) != '\0') && !((iq == 0) && isspace((int)c))) {		\
 	if ((c) == '"') {						\
 	    iq = !iq;							\
 	} else if (((c) == '\\') && (*(ptr) == '"')) {			\
@@ -841,7 +841,7 @@ void	areads_relbuf(int fd);
 
 #define	copy_string(ptr,c,f,l,fp) do {					\
     (fp) = (f);								\
-    while((c) != '\0' && !isspace(c)) {					\
+    while((c) != '\0' && !isspace((int)c)) {				\
 	if((fp) >= (f) + (l) - 1) {					\
 	    *(fp) = '\0';						\
 	    (fp) = NULL;						\
@@ -857,7 +857,7 @@ void	areads_relbuf(int fd);
 
 #define	copy_string_cs(ptr,c,f,l,fp) do {				\
     (fp) = (f);								\
-    while((c) != '\0' && (c) != '#' && !isspace(c)) {			\
+    while((c) != '\0' && (c) != '#' && !isspace((int)c)) {		\
 	if((fp) >= (f) + (l) - 1) {					\
 	    *(fp) = '\0';						\
 	    (fp) = NULL;						\
